test(kubernetes): add rendered manifest tests to AbstractInstallKT
Add shouldInstallFromRenderedManifest and shouldInstallFromCrdsOnlyThenOperator
tests to AbstractInstallKT so they run for all Kubernetes environments
(minikube, kind, etc).

Keep RenderedManifestKT for non-deployment validation tests
(shouldContainAllExpectedResources, shouldCrdsOnlyContainOnlyCrds,
shouldHaveNoUnsubstitutedVariables) which verify manifest structure
without requiring cluster deployment.
