+++ kroxylicious-runtime/javac.20260620_224615.args
@@ -1,0 +1,1 @@
+@/private/var/folders/5s/v5576rp175bf4vrf15mgvffm0000gn/T/org.codehaus.plexus.compiler.javac.JavacCompiler10184314390904994972arguments
